refactor: doing some refactoring on code #6
- Applying Abstraction on fade scripts
