+++ DAWN/Assets/Scripts/Cutscene/FadeBuilding.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 92a83055ee4f32648b81797c8d854179
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
