+++ DAWN/Assets/Scripts/Cutscene/FadeScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a001579ca954d9345927ec67ff11547e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
